requirements.txt
opencv-python
numpy
requests
torch
ultralytics
Pillow